+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 2.1.0 (2017-xx-xx)
 ------------------
 
+- Content disposition with semicolon in filename #917
 
 
 2.0.0 (2017-03-20)